:up: parse関数をファイル分割
